Separate  SeqAny state
