+++ lib/wallet/cardano-wallet.cabal
@@ -249,9 +249,10 @@ library
     Cardano.Wallet.Address.States.Features
     Cardano.Wallet.Address.States.IsOwned
     Cardano.Wallet.Address.States.Test.State
-    
+
     Cardano.Wallet.Address.MaybeLight
     Cardano.Wallet.Address.Discovery.RandomAny
+    Cardano.Wallet.Address.Discovery.SequentialAny
     Cardano.Wallet.Byron.Compatibility
     Cardano.Wallet.Checkpoints
     Cardano.Wallet.Checkpoints.Policy
