+++ lib/wallet/src/Cardano/Wallet/Address/Discovery/Sequential.hs
@@ -62,8 +62,6 @@ module Cardano.Wallet.Address.Discovery.Sequential
     , mkSeqStateFromAccountXPub
     , isOwned
 
-    -- ** Benchmarking
-    , SeqAnyState (..)
     ) where
 
 import Prelude
@@ -114,10 +112,6 @@ import Cardano.Wallet.Primitive.Passphrase
     ( Passphrase )
 import Cardano.Wallet.Primitive.Types.Address
     ( Address (..), AddressState (..) )
-import Cardano.Wallet.Primitive.Types.RewardAccount
-    ( RewardAccount )
-import Cardano.Wallet.Shelley.Compatibility.Ledger
-    ( toLedger )
 import Cardano.Wallet.TypeLevel
     ( Excluding )
 import Codec.Binary.Encoding
@@ -130,8 +124,6 @@ import Control.Monad
     ( unless )
 import Data.Bifunctor
     ( first )
-import Data.Digest.CRC32
-    ( crc32 )
 import Data.Kind
     ( Type )
 import Data.List.NonEmpty
@@ -150,13 +142,10 @@ import Fmt
     ( Buildable (..), blockListF', hexF, indentF, prefixF, suffixF )
 import GHC.Generics
     ( Generic )
-import GHC.TypeLits
-    ( KnownNat, Nat, natVal )
 import Type.Reflection
     ( Typeable )
 
 import qualified Cardano.Wallet.Address.Pool as AddressPool
-import qualified Cardano.Wallet.Write.UTxOAssumptions as UTxOAssumptions
 import qualified Data.List as L
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
@@ -603,83 +592,3 @@ instance (PaymentAddress k 'CredFromKeyK, HasSNetworkId n)
 
 instance GetAccount (SeqState n k) k where
     getAccount = accountXPub
-
-{-------------------------------------------------------------------------------
-    SeqAnyState
-
-    For benchmarking and testing arbitrary large sequential wallets.
--------------------------------------------------------------------------------}
-
--- | An "unsound" alternative that can be used for benchmarking and stress
--- testing. It re-uses the same underlying structure as the `SeqState` but
--- it discovers addresses based on an arbitrary ratio instead of respecting
--- BIP-44 discovery.
---
--- The proportion is stored as a type-level parameter so that we don't have to
--- alter the database schema to store it. It simply exists and depends on the
--- caller creating the wallet to define it.
-newtype SeqAnyState (network :: NetworkDiscriminant) key (p :: Nat) =
-    SeqAnyState { innerState :: SeqState network key }
-    deriving (Generic)
-
-deriving instance
-    ( Show (k 'AccountK XPub)
-    , Show (k 'CredFromKeyK XPub)
-    , Show (k 'PolicyK XPub)
-    , Show (KeyFingerprint "payment" k)
-    )
-    => Show (SeqAnyState n k p)
-
-instance
-    ( NFData (k 'AccountK XPub)
-    , NFData (k 'CredFromKeyK XPub)
-    , NFData (k 'PolicyK XPub)
-    , NFData (KeyFingerprint "payment" k)
-    )
-    => NFData (SeqAnyState n k p)
-
-instance KnownNat p => IsOurs (SeqAnyState n k p) Address where
-    isOurs (Address bytes) st@(SeqAnyState inner)
-        | crc32 bytes < p && correctAddressType =
-            let
-                pool = getPool $ externalPool inner
-                ix = toEnum $ AddressPool.size pool - AddressPool.gap pool
-                addr = AddressPool.addressFromIx pool ix
-                pool' = AddressPool.update addr pool
-                path = DerivationIndex (getIndex ix) :| []
-            in
-                ( Just path
-                , SeqAnyState $ inner{externalPool = SeqAddressPool pool'}
-                )
-        | otherwise =
-            (Nothing, st)
-      where
-        p = floor (double sup * double (natVal (Proxy @p)) / 10_000)
-          where
-            sup = maxBound :: Word32
-
-        double :: Integral a => a -> Double
-        double = fromIntegral
-
-        correctAddressType =
-            UTxOAssumptions.validateAddress
-                UTxOAssumptions.AllKeyPaymentCredentials
-                (toLedger $ Address bytes)
-
-instance IsOurs (SeqAnyState n k p) RewardAccount where
-    isOurs _account state = (Nothing, state)
-
-instance
-    ( SoftDerivation k
-    , AddressCredential k ~ 'CredFromKeyK
-    ) => GenChange (SeqAnyState n k p)
-  where
-    type ArgGenChange (SeqAnyState n k p) = ArgGenChange (SeqState n k)
-    genChange a (SeqAnyState s) = SeqAnyState <$> genChange a s
-
-instance SupportsDiscovery n k => CompareDiscovery (SeqAnyState n k p) where
-    compareDiscovery (SeqAnyState s) = compareDiscovery s
-
-instance (PaymentAddress k 'CredFromKeyK, HasSNetworkId n)
-    => KnownAddresses (SeqAnyState n k p) where
-    knownAddresses (SeqAnyState s) = knownAddresses s
