+++ lib/wallet/src/Cardano/Wallet/Address/States/Families.hs
@@ -12,7 +12,9 @@ import Cardano.Wallet.Address.Discovery.Random
 import Cardano.Wallet.Address.Discovery.RandomAny
     ( RndAnyState )
 import Cardano.Wallet.Address.Discovery.Sequential
-    ( SeqAnyState, SeqState )
+    ( SeqState )
+import Cardano.Wallet.Address.Discovery.SequentialAny
+    ( SeqAnyState )
 import Cardano.Wallet.Address.Discovery.Shared
     ( SharedState )
 import Cardano.Wallet.Address.States.Test.State
