+++ lib/wallet/src/Cardano/Wallet/Flavor.hs
@@ -53,7 +53,9 @@ import Cardano.Wallet.Address.Discovery.Random
 import Cardano.Wallet.Address.Discovery.RandomAny
     ( RndAnyState )
 import Cardano.Wallet.Address.Discovery.Sequential
-    ( SeqAnyState, SeqState )
+    ( SeqState )
+import Cardano.Wallet.Address.Discovery.SequentialAny
+    ( SeqAnyState )
 import Cardano.Wallet.Address.Discovery.Shared
     ( SharedState (..) )
 import Cardano.Wallet.Address.States.Families
